+++ changelog.md
@@ -56,7 +56,7 @@
 * `AssignToFeeders`/`AssignToLvFeeders` now finds back-fed equipment correctly
 * `AssignToFeeders` and `AssignToLvFeeders` will now associate `PowerElectronicUnits` with their `powerElectronicsConnection` `Feeder`/`LvFeeder`.
 * Phases are now correctly assigned to the LV side of an LV2 transformer that is in parallel with a previously energised LV1 transformer.
-
+* Added missing default network state operators (NORMAL) if you are calling `SetDirection.run` directly.
 
 ### Notes
 * None.